Refactoring code out of main, in progress
